bind, mount: stop returning mount id
The mount ID is a sequence number in a 32 bit integer, which means that
it can't be unique. This is largely harmless, because there is no way to
use the mount id, beyond checking if it's negative.

However, there's no overflow check, so the mount ID can wrap negative,
which will break error checks on mount calls.

Because it's useless, let's just stop returning it.