dtracy: correctly look up aggregate keys
the ANode struct contains a variable sized buffer for the key,
however when we created and copied the aggregate, we used struct
assignment to initialize the node. the struct assignment only
knows about the fixed size portion of the struct, and did not
copy the key, which mean that key lookups would fail, and we would
insert a new value into the aggregation every time, both leaking
the memory and producng incorrect results.