dtracy: make timer probes run in interrupt context
When probing a timer, we were running in our own kproc,
and not in an interrupt context, which meant that we didn't
have any access to anything worth sampling, so we didn't
give any data back.

This moves the probe to the hzclock interrupt, and returns
the pc in the probe.