window: add wsys 'none' attach to attempted wctl search
While window did work when run under riostart without $wctl,
the change window running from the plumber. It is typical
for lib/profile to contain:

plumber
rio -s riostart

Previously this worked because rio polluted the envgroup
with $wctl and $wsys, but without $wctl or /dev/wctl the
plumber's window fails to find rio. In favor of not breaking
this configuration we allow window to fallback to checking $wsys
and attaching to the none window. This should keep window working
in all cases in which the wctl pollution worked before.