cc: take our pickle out of the peephole
The -P flag is used to debug the peephole optimizer. The -Z flag is
used to output pickling code for various types. Don't confuse the two.

The pickling code was added in the third edition. It mistakenly reused
the -P flag which was later partially corrected to use the -Z flag in
2004. That change updated lex.c but missed the code in pickle.c.