ktrans: revisit grammer handling and do some spring cleaning
Revisiting the man page example and README from the
original ktrans there was some descripency on how
to handle punction marks and special characters.

Notably 。(.) and 、(,) need special casing. If we
have anything in the buffer we need to tack on
these characters as an Okuri tail and avoid the actual
punction from entering the Kanji buffer.

Newlines were previously modified to be taken as a completion
if there was runes in the buffer. This has been backed out,
instead Shift + Space can serve this role via kbmap should they
prefer (and as is done in the jp kbmap). Instead we treat newlines
as hints to reset the buffer.

There was also a bug in where after cycling through all options the
original hiragana was not printed back. This has been corrected.

The max number of candidates has been bumped to 32 and moved to an enum.
This does nearly double our resident memory size, but we reguarly had
matches exceed this limit. A better solution is slated.

The man page now makes an attempt to explain the rules around Okuri and
Joshi input modes.