kernel: fix freeb() for custom pools
Some ethernet drivers like ethervt6105M maintain
a custom pool of Blocks using Block.free
callback.

To maintain that Block.list pointer is
nil when reused, we have to clear it before
returning it to the custom pool as the custom
pool code is not aware of it.

Also, poison Block.list pointer before free().