patch: improve hunk search
Keep track of last hunk offset so we are more likely to find future
hunks.

Start search within the range of lines where we could possibly find
the hunk.  This fixes a bug where if the file shrunk since the patch
was generated, and we start the search past the end of the file and
immediately abort the search.

Add an extra offset for the end of the file to the lines array so that
lines[nlines] is the length of the file.  This way, when we start the
search at line nlines-oldcnt, we don't access past the end of the array
if oldcnt==0.

When we find a hunk, set lastln to ln + h->oldcnt; we can't apply
another hunk that starts before the previous hunk ends.