reform: only override console=0 when no plan9.ini is passed
The kernel used to always set console=0, which as
usefull during bring-up, but makes it impossible
to use the uart for other purposes.

We now have the ability to pass plan9.ini using
the u-boot script, so add the console=0 line there.

To make debugging easy, we still apply
console=0 if no plan9.ini has been passed.