ircrc: fix race with pipe causing hangup
If the work function opens and closes $netdir/data for an echo
redirection *before* the asynchronous cat process opens it,
aux/trampoline (or tlsclient) will receive a hangup, terminating the
connection.

To fix this, redirect work's stdout to $netdir/data so that it remains
open, and adjust the echo commands accordingly.