nusb/audio: only consider data endpoints when setting up
An async iso endpoint has a data endpoint and feedback endpoint, and
both get mapped to the same slot in d->usb->ep. However, when setting
up audio streams, we are only interested in the data endpoints. If
the feedback endpoint appears first in the list, we end up trying to
use feedback endpoints for audio data and ignore the data endpoints.