9boot: delay GetMemoryMap until right before ExitBootServices (thanks michael forney!)
Michael Forney reported the following:

The UEFI specification suggests calling GetMemoryMap immediately
before ExitBootServices to ensure that the loader has the current
memory map. The firmware is able to enforce this since we have to
pass a "MapKey" from the last GetMemoryMap to ExitBootServices.

The T14 AMD gen 1 firmware is quite strict about this, and even a call
to OutputString will invalidate the MapKey. This causes
ExitBootServices to fail, and we enter 9front with boot services still
running. This causes all sorts of problems including leaving the
IOMMU enabled, breaking 9front's PCI drivers.

---

Michael sent a patch that moves memconf() to unload(), but i prefer
to make this explicit by calling memconf() before and keep unload() as
just that, not touching the configuration.