devsrv: Various fixes.
Make global variables static to avoid symbol clashes.

Order Link fields more compactly for 64-bit archs.

Add a common freelink() function for freeing Boards and Srvs.

Chain all Boards in a circular doubly-linked list,
so srvrename() can actually change the owners of *all*
the boards instead of just the root. We cannot use
recursion here as that could potentially blow the
kernel stack.

srvname() was also only processing the root.

Instead, we add a srvname field to Chan and set it to
the original path of the srv file that was used to post
the channel.

Call openmode() at the beginning of srvopen().
The reason is if omode has invalid bits it can error
and leave stuff in a inconsistent state.

Prevent srvwstat() to rename a file to "clone", it is
reserved.

Get rid of "lease expired" error message, just use
Eshutdown.