nusb/usbd: only fetch first 8 bytes of device descriptor for getmaxpkt() (thanks k0ga)
The first usb transaction we run on a device is reading the
device descriptor, but todo this we need to know the maximum
packet size on the control endpoint.

But the packet size itself is stored in the device descriptor
which needs to be read over the control endpoint.

We used to fetch up to 64 bytes of device descriptor, and
if that fails assume some default values.

But this seems to cause errors down the line for some devices
like k0gas usb keyboard.

The new way is to read *ONLY* the first 8 bytes (that contain
the bMaxPktSize0 field) and fetch the full device descriptor
later once we have set the correct packet size.