kbmap, kbremap: open kbmap file with OTRUNC and bugfixes
Now that kbdfs supports reseting the map to
the default itself we no longer need to
manually load the ascii maps.

This also fixes a 'bug in the making' in
the buffer read logic that would cause it
to cut lines in half when writing to the
kbmap file. All of our kbmaps are under
8k so this never showed up.

Also fix a small bug in kbremap where it
was not writing out the last null when writing out
kbd events.