sdnvme: add smart file for full smart info
Instead of cramming everything to the ctl file,
provide a separate file with the SMART/health information
using the format that matches the NVMe specification closely.

Also, the command to read the health information log page
can be namespace specific as indicated by bit 0 of LPA in
controller identification information.