ndb/dns: hack for the caching server for delegated nameservers
This is a hack for a delegation to a nameservers that
simply do not answer with authorative replies.

This results in us asking some higher up nameserver
when no answer is returned, but because the ns record
is itself already authorative, it wont be able to update
any new info (rrattach()).

Instead, we cut the search short if the nameserver
ns record is set as authorative in ndb (with soa=).

it is a unfortunate hack because this behaviour cannot
be implemented for non-caching server as it does not
load the authorative areas (inmyarea()).