kernel: make walk/open errors more consistent with userspace
As is the error for an open would be in the form:
	'filename' does not exist
This changes it to be of the form:
	file does not exist: 'filename'

This brings it more in line with what could
be expected from a userspace filesystem.
Existing code, perhaps due to this discrepancy,
all uses strstr to check for this error. They can
now instead check for explicitly the error as a prefix.
This would avoid issues with potentially maliciously
named files such as "does not exist".

Likewise this also helps filesystems such as exportfs,
which spit back the errstr from open(2) directly as Rerror
to their clients.