patch: fix offsets in reverse mode
we would not swap all of the fields in the forward
and reverse modes for patch, which means that we
may accidentally make a hunk unfindable if the line
numbers matcch up correctly.

This does the reverse in a less ad-hoc way, just
swapping the forward and reverse hunks once in
one place. It also swaps all of them.