nusbrc: use exponential backoff and timeout usb enumeration
nusbrc used to wait forever until initial device
enumeration was completed. the reason is mainly
so that usb stoage or ethernet devices can be
used as boot media.

but this can result in a system getting completely stuck
on boot if usb enumeration gets stuck. (this really
depends on the device driver if some timeouts
have been implemented during attach).

as a compromise, we will wait for usb enumeration
to complete using exponential back-off with a maximum
total wait time of ~10 seconds.

typical times for single device initial enumeration is
0.4 to 0.6 seconds.

largest case i measured with 20 random devices was
7 seconds.

so we will probably be safe with the 10 second timeout.