7l: use wider INITTEXT and INITDAT
Currently we use a 32-bit value for these values which means
that large values passed to -T will be truncated and sign-extended,
rather than stored verbatim. This means load addresses in between
low 2GB and high 2GB will not be respected.

It appears this has been fixed in 6l already, so we can leave it alone.